Tests didn't run with this util file
