+++ api_tests/lib_sdk/actors/actor.ts
@@ -13,6 +13,8 @@ import { sleep } from "../utils";
 import { Wallet, Wallets } from "../wallets";
 import { Actors } from "./index";
 import { HarnessGlobal } from "../../lib/util";
+import {Entity} from "../../gen/siren";
+import {SwapDetails} from "comit-sdk/dist/src/cnd";
 
 declare var global: HarnessGlobal;
 
@@ -685,6 +687,40 @@ export class Actor {
     public cndHttpApiUrl() {
         const cndSocket = this.cndInstance.getConfigFile().http_api.socket;
         return `http://${cndSocket.address}:${cndSocket.port}`;
+    }
+
+    public async pollCndUntil(
+        location: string,
+        predicate: (body: Entity) => boolean
+    ): Promise<Entity> {
+        const response = await this.cnd.fetch(location);
+
+        expect(response).to.have.status(200);
+
+        if (predicate(response.data)) {
+            return response.data;
+        } else {
+            await sleep(500);
+
+            return this.pollCndUntil(location, predicate);
+        }
+    }
+
+    public async pollSwapDetails(
+        swapUrl: string,
+        iteration: number = 0
+    ): Promise<SwapDetails> {
+        if (iteration > 5) {
+            throw new Error(`Could not retrieve Swap ${swapUrl}`);
+        }
+        iteration++;
+
+        try {
+            return (await this.cnd.fetch<SwapDetails>(swapUrl)).data;
+        } catch (error) {
+            await sleep(1000);
+            return await this.pollSwapDetails(swapUrl, iteration);
+        }
     }
 }
