+++ api_tests/lib_sdk/utils.ts
@@ -1,6 +1,10 @@
-export async function sleep(ms: number) {
+import {ethers} from "ethers";
+import {Actor} from "./actors/actor";
+import {SwapRequest} from "comit-sdk";
+
+export async function sleep(time: number) {
     return new Promise(res => {
-        setTimeout(res, ms);
+        setTimeout(res, time);
     });
 }
 
@@ -15,4 +19,130 @@ export async function timeout<T>(ms: number, promise: Promise<T>): Promise<T> {
 
     // Returns a race between our timeout and the passed in promise
     return Promise.race([promise, timeout]);
+}
+
+export const DEFAULT_ALPHA = {
+    ledger: {
+        name: "bitcoin",
+        network: "regtest",
+    },
+    asset: {
+        name: "bitcoin",
+        quantity: {
+            bob: "100000000",
+            charlie: "200000000",
+            reasonable: "100000000",
+            stingy: "100",
+        },
+    },
+    expiry: new Date("2080-06-11T23:00:00Z").getTime() / 1000,
+};
+
+const DEFAULT_BETA = {
+    ledger: {
+        name: "ethereum",
+        chain_id: 17,
+    },
+    asset: {
+        name: "ether",
+        quantity: {
+            bob: ethers.utils.parseEther("10").toString(),
+            charlie: ethers.utils.parseEther("20").toString(),
+        },
+    },
+    expiry: new Date("2080-06-11T13:00:00Z").getTime() / 1000,
+};
+export async function createDefaultSwapRequest(counterParty: Actor) {
+    const swapRequest: SwapRequest = {
+        alpha_ledger: {
+            name: DEFAULT_ALPHA.ledger.name,
+            network: DEFAULT_ALPHA.ledger.network,
+        },
+        beta_ledger: {
+            name: DEFAULT_BETA.ledger.name,
+            chain_id: DEFAULT_BETA.ledger.chain_id,
+        },
+        alpha_asset: {
+            name: DEFAULT_ALPHA.asset.name,
+            quantity: DEFAULT_ALPHA.asset.quantity.bob,
+        },
+        beta_asset: {
+            name: DEFAULT_BETA.asset.name,
+            quantity: DEFAULT_BETA.asset.quantity.bob,
+        },
+        beta_ledger_redeem_identity:
+            "0x00a329c0648769a73afac7f9381e08fb43dbea72",
+        alpha_expiry: DEFAULT_ALPHA.expiry,
+        beta_expiry: DEFAULT_BETA.expiry,
+        peer: {
+            peer_id: await counterParty.cnd.getPeerId(),
+            address_hint: await counterParty.cnd
+                .getPeerListenAddresses()
+                .then(addresses => addresses[0]),
+        },
+    };
+    return swapRequest;
+}
+
+export const DEFAULT_ALPHA = {
+    ledger: {
+        name: "bitcoin",
+        network: "regtest",
+    },
+    asset: {
+        name: "bitcoin",
+        quantity: {
+            bob: "100000000",
+            charlie: "200000000",
+            reasonable: "100000000",
+            stingy: "100",
+        },
+    },
+    expiry: new Date("2080-06-11T23:00:00Z").getTime() / 1000,
+};
+
+const DEFAULT_BETA = {
+    ledger: {
+        name: "ethereum",
+        chain_id: 17,
+    },
+    asset: {
+        name: "ether",
+        quantity: {
+            bob: ethers.utils.parseEther("10").toString(),
+            charlie: ethers.utils.parseEther("20").toString(),
+        },
+    },
+    expiry: new Date("2080-06-11T13:00:00Z").getTime() / 1000,
+};
+export async function createDefaultSwapRequest(counterParty: Actor) {
+    const swapRequest: SwapRequest = {
+        alpha_ledger: {
+            name: DEFAULT_ALPHA.ledger.name,
+            network: DEFAULT_ALPHA.ledger.network,
+        },
+        beta_ledger: {
+            name: DEFAULT_BETA.ledger.name,
+            chain_id: DEFAULT_BETA.ledger.chain_id,
+        },
+        alpha_asset: {
+            name: DEFAULT_ALPHA.asset.name,
+            quantity: DEFAULT_ALPHA.asset.quantity.bob,
+        },
+        beta_asset: {
+            name: DEFAULT_BETA.asset.name,
+            quantity: DEFAULT_BETA.asset.quantity.bob,
+        },
+        beta_ledger_redeem_identity:
+            "0x00a329c0648769a73afac7f9381e08fb43dbea72",
+        alpha_expiry: DEFAULT_ALPHA.expiry,
+        beta_expiry: DEFAULT_BETA.expiry,
+        peer: {
+            peer_id: await counterParty.cnd.getPeerId(),
+            address_hint: await counterParty.cnd
+                .getPeerListenAddresses()
+                .then(addresses => addresses[0]),
+        },
+    };
+    return swapRequest;
 }
